+++ packages/node-core/src/configure/NodeConfig.ts
@@ -30,7 +30,6 @@ export interface IConfig {
   readonly logLevel: LevelWithSilent;
   readonly queryLimit: number;
   readonly indexCountLimit: number;
-  readonly timestampField: boolean;
   readonly proofOfIndex: boolean;
   readonly ipfs?: string;
   readonly dictionaryTimeout: number;
@@ -68,7 +67,6 @@ const DEFAULT_CONFIG = {
   debug: undefined,
   queryLimit: 100,
   indexCountLimit: 10,
-  timestampField: true,
   proofOfIndex: false,
   dictionaryTimeout: 30,
   dictionaryQuerySize: 10000,
@@ -215,10 +213,6 @@ export class NodeConfig<C extends IConfig = IConfig> implements IConfig {
 
   get indexCountLimit(): number {
     return this._config.indexCountLimit;
-  }
-
-  get timestampField(): boolean {
-    return this._config.timestampField;
   }
 
   get proofOfIndex(): boolean {